requirements.txt
streamlit
google-cloud-storage
google-cloud-documentai
google-cloud-aiplatform
vertexai
PyPDF2